refactor ModelFlatFile
